requirements_local.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
# greenlet foi removido pois requer compilacao e nao e essencial para SQLAlchemy com SQLite/PostgreSQL sync.
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.3
python-dateutil==2.9.0.post0
pytz==2024.1
SQLAlchemy==2.0.31
typing-extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
WTForms==3.1.2
waitress==3.0.0
colorama==0.4.6
six==1.16.0
pandas
python-dotenv

pandas
python-dotenv
